Ensure we're not repeatedly resampling traces
Traces that were unable to be written to the collector were
being repeatedly reprocessed and resampled. As rejected samples
are immediately removed, this means that all samples are likely
to be rejected after multiple resampling.

Preserve the original sampling decision to avoid resampling,
and add a timeout provision if the collector isn't accepting traces
for whatever reason